align-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
numpy==1.26.4
scipy==1.13.1
librosa==0.10.2.post1
soundfile==0.12.1
music21==9.3.0
PyGuitarPro==0.9.3
pydantic==2.10.3
